1.增加redis配置工具类； 2.增加oss连接、配置工具类； 3.增加pc端用户增加、登录、登出接口

+++ src/main/java/com/qyt/qytbackend/mapper/UserInfoMapper.java
@@ -12,7 +12,7 @@ public interface UserInfoMapper {
 
     int update(UserInfo userInfo);
 
-    UserInfo selectById(@Param("id") Integer id);
+    UserInfo selectById(@Param("id") Long id);
 
     UserInfo selectByUsername(@Param("username") String username);
